requirements.txt
boto3==1.34.76
pyjwt
fastapi
starlette
uvicorn
loguru